% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksj_bind.R
\name{ksj_bind}
\alias{ksj_bind}
\title{Bind KSJ files in directory}
\usage{
ksj_bind(dir, identifier, regexp = NULL, translate = NULL)
}
\arguments{
\item{dir}{path}

\item{identifier}{Data identifier. (e.g. "N02")}

\item{regexp}{A regular expression passed on target directory.}

\item{...}{Additional arguments to parser function}
}
\description{
Bind KSJ files in directory
}
