% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_isj.R
\name{read_isj}
\alias{read_isj}
\title{Read ISJ format data}
\usage{
read_isj(
  path = NULL,
  .area_code = NULL,
  .fiscal_year = NULL,
  .pos_level = NULL,
  .download = FALSE,
  return_class = NULL
)
}
\arguments{
\item{path}{Input file path (`csv`)}

\item{.area_code}{Area code. 5 digits numeric character.
E.g. Chiba prefecture ("12000"), Ichikawa city in Chiba ("12033")}

\item{.fiscal_year}{Japanese imperial fiscal year as Kanji.}

\item{.pos_level}{E.g. 0 is to street level 3. 1 is to street level 1.}

\item{.download}{If `TRUE`, download file to working directory.}

\item{return_class}{Select "tbl_df" (data.frame) or "data.table"}
}
\description{
Read ISJ format data
}
\seealso{
\url{http://nlftp.mlit.go.jp/isj/about_api.html}
}
