% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ksj_a30a5.R
\name{read_ksj_a30a5}
\alias{read_ksj_a30a5}
\title{Kokudosuuchi A30a5 parser}
\usage{
read_ksj_a30a5(path = NULL, .meshcode = NULL, .download = FALSE)
}
\arguments{
\item{path}{Input file path (`.shp` or `.geojson`)}

\item{.meshcode}{Specific 80km meshcode}

\item{.download}{If `true`, the specified file is downloaded to the current working directory.}
}
\description{
If there is no local file, specify the meshcode to download.
}
