% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ksj_c02.R
\name{read_ksj_c02}
\alias{read_ksj_c02}
\title{Kokudosuuchi C02 parser}
\usage{
read_ksj_c02(path = NULL, translate = "jp", .year = NULL, .download = FALSE)
}
\arguments{
\item{path}{Input file path (`.shp` or `.geojson`)}

\item{translate}{Translate the variable name into the specified language.
You can choose either Original (`raw`) and Japanese (`jp`) or English (`en`).}

\item{.year}{Specific year}

\item{.download}{If `true`, the specified file is downloaded to the current working directory.}
}
\description{
If there is no local file, specify the year to download.
}
