% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ksj_c23.R
\name{read_ksj_c23}
\alias{read_ksj_c23}
\title{Kokudosuuchi C23 parser}
\usage{
read_ksj_c23(path = NULL, .pref_code = NULL, .download = FALSE)
}
\arguments{
\item{path}{Input file path (`.shp` or `.geojson`)}

\item{.pref_code}{Specific prefecture code}

\item{.download}{If `true`, the specified file is downloaded to the current working directory.}
}
\description{
If there is no local file, specify prefecture code to download.
}
