% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ksj_l03a.R
\name{read_ksj_l03a}
\alias{read_ksj_l03a}
\title{Kokudosuuchi L03a parser}
\usage{
read_ksj_l03a(path = NULL, .year = NULL, .meshcode = NULL, .download = FALSE)
}
\arguments{
\item{path}{Input file path (`.shp` or `.geojson`)}

\item{.year}{Specific year}

\item{.meshcode}{Specific 80km meshcode}

\item{.download}{If `true`, the specified file is downloaded to the current working directory.}
}
\description{
If there is no local file, specify year and meshcode to download.
}
