% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ksj_l03b.R
\name{read_ksj_l03b}
\alias{read_ksj_l03b}
\title{Kokudosuuchi L03-b parser}
\usage{
read_ksj_l03b(
  path = NULL,
  .year,
  .meshcode,
  .datum,
  .download = FALSE,
  .translate = "jp"
)
}
\arguments{
\item{path}{Input file path (`.shp` or `.geojson`)}

\item{.year}{Specific year}

\item{.meshcode}{Specific 80km meshcode}

\item{.datum}{"jgd" or "tky"}

\item{.download}{If `true`, the specified file is downloaded to the current working directory.}

\item{.translate}{Translate the variable name into the specified language.
You can choose either Japanese (`jp`) or English (`en`).}
}
\description{
If there is no local file, specify the meshcode to download.
}
