% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ksj_n03.R
\name{read_ksj_n02}
\alias{read_ksj_n02}
\title{Kokudosuuchi N02 parser}
\usage{
read_ksj_n02(
  path = NULL,
  .year = NULL,
  .download = FALSE,
  .type = c("railroadsection", "station")
)
}
\arguments{
\item{path}{Input file path (`.shp` or `.geojson`)}

\item{.year}{Specific year}

\item{.download}{If `true`, the specified file is downloaded to the current working directory.}

\item{.type}{The type of file. "Railway section (railroadsection)" (default) or "station"}
}
\description{
If there is no local file, specify the year to download.
}
