% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ksj_p09.R
\name{read_ksj_p09}
\alias{read_ksj_p09}
\title{Kokudosuuchi P09 parser}
\usage{
read_ksj_p09(path, .meshcode = NULL, .download = FALSE)
}
\arguments{
\item{path}{Input file path (`.shp` or `.geojson`)}

\item{.meshcode}{Specific 80km meshcode}

\item{.download}{If `true`, the specified file is downloaded to the current working directory.}
}
\description{
Kokudosuuchi P09 parser
}
