% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ksj_p12.R
\name{read_ksj_p12}
\alias{read_ksj_p12}
\title{Kokudosuuchi P12 parser}
\usage{
read_ksj_p12(path = NULL, .pref_code = NULL, .download = FALSE,
  .type = c("point", "line", "polygon"))
}
\arguments{
\item{path}{Input file path (`.shp` or `.geojson`)}

\item{.pref_code}{Specific prefecture code}

\item{.download}{If `true`, the specified file is downloaded to the current working directory.}

\item{.type}{File type. When downloading a file, select the type of data to use, either "point", "line" or "polygon".}
}
\description{
If there is no local file, specify prefecture code to download.
}
