% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ksj_p34.R
\name{read_ksj_p34}
\alias{read_ksj_p34}
\title{Kokudosuuchi P34 parser}
\usage{
read_ksj_p34(path = NULL, .pref_code = NULL, .download = FALSE)
}
\arguments{
\item{path}{Input file path (`.shp` or `.geojson`)}

\item{.pref_code}{Specific prefecture code}

\item{.download}{If `true`, the specified file is downloaded to the current working directory.}
}
\description{
If there is no local file, specify the year and pref_code to download.
}
