% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ksj_w05.R
\name{read_ksj_w05}
\alias{read_ksj_w05}
\title{Kokudosuuchi W05 parser}
\usage{
read_ksj_w05(path = NULL, .pref_code = NULL, .download = FALSE,
  .type = c("stream", "node"))
}
\arguments{
\item{path}{Input file path (`.shp` or `.geojson`)}

\item{.pref_code}{Specific prefecture code}

\item{.download}{If `true`, the specified file is downloaded to the current working directory.}

\item{.type}{File type.}
}
\description{
If there is no local file, specify the prefecture to download.
}
