% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ksj_a16.R
\name{read_ksj_a16}
\alias{read_ksj_a16}
\title{Kokudosuuchi A10 parser}
\usage{
read_ksj_a16(
  path = NULL,
  .pref_code = NULL,
  .year = NULL,
  .download = TRUE,
  .translate = NULL
)
}
\arguments{
\item{path}{Input file path (`.shp` or `.geojson`)}

\item{.pref_code}{Specific prefecture code}

\item{.year}{Specific year}

\item{.download}{If `true`, the specified file is downloaded to the current working directory.}

\item{.translate}{Translate the variable name into the specified language.
You can choose either Japanese (`jp`) or English (`en`).}
}
\description{
If there is no local file, specify the year and pref_code to download.
}
