% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ksj_l03b.R
\name{read_ksj_l03b}
\alias{read_ksj_l03b}
\title{Kokudosuuchi L03-b parser}
\usage{
read_ksj_l03b(
  path = NULL,
  translate = "jp",
  .year,
  .meshcode,
  .datum,
  .download = FALSE
)
}
\arguments{
\item{path}{Input file path (`.shp` or `.geojson`)}

\item{translate}{Translate the variable name into the specified language.
You can choose either Original (`raw`) and Japanese (`jp`) or English (`en`).}

\item{.year}{Specific year}

\item{.meshcode}{Specific 80km meshcode}

\item{.datum}{"jgd" or "tky"}

\item{.download}{If `true`, the specified file is downloaded to the current working directory.}
}
\description{
If there is no local file, specify the meshcode to download.
}
