% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ksj_p23.R
\name{read_ksj_p23}
\alias{read_ksj_p23}
\title{Kokudosuuchi P23 parser}
\usage{
read_ksj_p23(
  path = NULL,
  translate = "jp",
  .pref_code = NULL,
  .download = FALSE,
  .type
)
}
\arguments{
\item{path}{Input file path (`.shp` or `.geojson`)}

\item{translate}{Translate the variable name into the specified language.
You can choose either Original (`raw`) and Japanese (`jp`) or English (`en`).}

\item{.pref_code}{Specific prefecture code}

\item{.download}{If `true`, the specified file is downloaded to the current working directory.}

\item{.type}{File type. When downloading a file, select the type of data to use, either "point" or "line".}
}
\description{
If there is no local file, specify prefecture code to download.
}
